RavenDB-18561 - Catch the Script raised exception in TryResolveConflictByScriptInternal
