+++ src/Raven.Server/Documents/Replication/ConflictManager.cs
@@ -5,8 +5,6 @@ using System.Linq;
 using Raven.Client.Extensions;
 using Raven.Client.ServerWide;
 using Raven.Server.Documents.Handlers;
-using Raven.Server.NotificationCenter.Notifications;
-using Raven.Server.NotificationCenter.Notifications.Details;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.Utils;
 using Sparrow.Json;
@@ -72,9 +70,9 @@ namespace Raven.Server.Documents.Replication
                 if (IsSameCollection(documentsContext, id, conflictedDoc.Collection))
                 {
                     if (TryResolveConflictByScript(
-                            documentsContext,
-                            conflictedDoc))
-                            return;
+                        documentsContext,
+                        conflictedDoc))
+                        return;
 
                     if (_database.ReplicationLoader.ConflictSolverConfig?.ResolveToLatest ?? true)
                     {
@@ -173,33 +171,15 @@ namespace Raven.Server.Documents.Replication
 
             conflictedDocs.Add(conflict.Clone());
             conflictedDocs.Sort((x, y) => string.Compare(x.ChangeVector, y.ChangeVector, StringComparison.Ordinal));
-            try
-            {
-                if (_conflictResolver.TryResolveConflictByScriptInternal(
-                        documentsContext,
-                        scriptResolver,
-                        conflictedDocs,
-                        documentsContext.GetLazyString(collection), out var resolved))
-                {
-                    _conflictResolver.PutResolvedDocument(documentsContext, resolved, resolvedToLatest: false, conflict);
-                    return true;
-                }
-            }
-            catch (Exception e)
-            {
-                var msg = $"Script failed to resolve the conflict in doc: {conflict.Id} because exception was raised in it.";
-                if (_log.IsInfoEnabled)
-                    _log.Info(msg, e);
 
-                var alert = AlertRaised.Create(
-                    _database.Name,
-                    "Conflict didn't resolved",
-                    msg,
-                    0,
-                    NotificationSeverity.Error,
-                    details: new ExceptionDetails(e));
-
-                _database.NotificationCenter.Add(alert);
+            if (_conflictResolver.TryResolveConflictByScriptInternal(
+                documentsContext,
+                scriptResolver,
+                conflictedDocs,
+                documentsContext.GetLazyString(collection), out var resolved))
+            {
+                _conflictResolver.PutResolvedDocument(documentsContext, resolved, resolvedToLatest: false, conflict);
+                return true;
             }
 
             return false;
